# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Foot Trading API",
    "version": "1.0.0",
    "description": "API documentation for Foot Trading project"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "summary": "Crear un nuevo usuario",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al crear el usuario"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Iniciar sesión",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nombre de usuario o contraseña incorrectos"
          },
          "500": {
            "description": "Error al iniciar sesión"
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "summary": "Obtener el perfil del usuario",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Nombre de usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil del usuario obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "birthDate": {
                      "type": "string",
                      "format": "date"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error al obtener el perfil del usuario"
          }
        }
      }
    },
    "/upload-product": {
      "post": {
        "summary": "Crear un nuevo producto",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "titulo": {
                    "type": "string"
                  },
                  "precio": {
                    "type": "number"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "tipo": {
                    "type": "string"
                  },
                  "talla": {
                    "type": "string"
                  },
                  "certificadoAutenticidad": {
                    "type": "boolean"
                  },
                  "categoria": {
                    "type": "string"
                  },
                  "imagenesAdjuntas": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al crear el producto"
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Obtener todos los productos",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener los productos"
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "summary": "Borrar un usuario por ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario borrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al borrar el usuario"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/products/{id}": {
      "delete": {
        "summary": "Borrar un producto por ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID del producto"
          }
        ],
        "responses": {
          "200": {
            "description": "Producto borrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al borrar el producto"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un producto por ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID del producto"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titulo": {
                    "type": "string"
                  },
                  "precio": {
                    "type": "number"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "tipo": {
                    "type": "string"
                  },
                  "talla": {
                    "type": "string"
                  },
                  "certificadoAutenticidad": {
                    "type": "boolean"
                  },
                  "categoria": {
                    "type": "string"
                  },
                  "imagenesAdjuntas": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al actualizar el producto"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
    "/create-payment-intent": {
      "post": {
        "summary": "Crear un PaymentIntent para Stripe",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartItems": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "titulo": {
                          "type": "string"
                        },
                        "precio": {
                          "type": "number"
                        },
                        "quantity": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PaymentIntent creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clientSecret": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al crear el PaymentIntent"
          }
        }
      }
    },
    "/payment-success": {
      "post": {
        "summary": "Confirmar el éxito del pago y eliminar productos comprados",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartItems": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Productos eliminados exitosamente"
          },
          "500": {
            "description": "Error al eliminar productos"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "required": [
          "titulo",
          "precio",
          "descripcion"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del producto"
          },
          "titulo": {
            "type": "string",
            "description": "Título del producto"
          },
          "precio": {
            "type": "number",
            "description": "Precio del producto"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción del producto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del producto"
          },
          "talla": {
            "type": "string",
            "description": "Talla del producto (solo para ropa)"
          },
          "certificadoAutenticidad": {
            "type": "boolean",
            "description": "Certificado de autenticidad (solo para cartas)"
          },
          "categoria": {
            "type": "string",
            "description": "Categoría del accesorio (solo para accesorios)",
            "enum": [
              "Balones",
              "Espinilleras",
              "Otro"
            ]
          },
          "imagenesAdjuntas": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "URLs de las imágenes adjuntas"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "username",
          "password",
          "email",
          "firstName",
          "lastName",
          "birthDate",
          "address"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del usuario"
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario"
          },
          "email": {
            "type": "string",
            "description": "Email del usuario"
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario"
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos del usuario"
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "description": "Fecha de nacimiento del usuario"
          },
          "address": {
            "type": "string",
            "description": "Dirección del usuario"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del usuario"
          }
        }
      }
    }
  },
  "tags": []
}